requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.8.0
psycopg2==2.9.1
python-dotenv==0.19.1
drf-yasg==1.20.0
